# app/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /api/corpus/:
    get:
      operationId: listCorpus
      description: Get all corpus. A corpus is a "category" that a profile belong to.
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?page=4
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?page=2
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Corpus'
          description: ''
      tags:
      - corpus
    post:
      operationId: createCorpus
      description: Add a corpus.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Corpus'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Corpus'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Corpus'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Corpus'
          description: ''
      tags:
      - corpus
...
components:
  schemas:
...
    Corpus:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        corpus_name:
          type: string
          maxLength: 1000
      required:
      - corpus_name
...